+++ src/pipeline/quorum_pipeline.py
@@ -18,7 +18,7 @@ import re
 import subprocess
 import pickle
 import tempfile
-
+import glob
 base_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 for directory_name in os.listdir(base_path):
     site.addsitedir(os.path.join(base_path, directory_name))
@@ -33,12 +33,14 @@ import nbayes
 import rforests
 import cdhit
 import fasta
+from hmmer import *
 from quorum import *
 
 class QuorumPipelineHandler(object):
     def __init__(self,
                  rootdir,
                  genome_files,
+                 six_frame_genomes,
                  intergenes,
                  annotated_genes,
                  bacteriocins,
@@ -58,6 +60,7 @@ class QuorumPipelineHandler(object):
         #Declare global vars
         self.rootdir            =     rootdir   			
         self.genome_files       =     genome_files
+        self.six_frame_genomes  =     six_frame_genomes
         self.bacteriocins   	=	  bacteriocins  		
         self.bacteriocin_radius =	  bacteriocin_radius	
         self.similarity	        =	  similarity			
@@ -88,19 +91,34 @@ class QuorumPipelineHandler(object):
         self.classifier_out = "%s/classify"%self.intermediate
         #Declare object handlers
         self.clusterer = None                       
-        self.textClassifier = None           
+        self.textClassifier = None
+        self.hmmers = []   
         self.nbpickle = "nb.zip"
         self.clusterpickle = "cluster.zip"
         self.textout = "text_out.txt"
         self.jobs  = []
         self.split_files = [] 
-        
+        self.classes = ["toxin","modifier","immunity","transport","regulator"]
+        self.class_files = ["%s/%s.fa"%(self.intermediate,s) for s in self.classes]
+        self.class_out = ["%s/%s.out"%(self.intermediate,s) for s in self.classes] 
+        self.batch_files = []
     def cleanup(self):
         print "Cleaning up"
         for job in self.jobs:
             job.erase_files()
         for fname in self.split_files:
-            os.remove(fname) 
+            if os.path.exists(fname): os.remove(fname)
+        for hmmer in self.hmmers:
+            hmmer.cleanUp() 
+        for i in xrange(len(self.class_files)):
+            if os.path.exists(self.class_files[i]): os.remove(self.class_files[i])
+        print glob.glob("tmp*")
+        for file in glob.glob("tmp*"):
+            os.remove(file)
+        for file in self.batch_files:
+            for extra in glob.glob("%s*"%file):
+                os.remove(extra)
+            
     """ Builds database such as the 
         intergenic database
         annotated genes database 
@@ -119,7 +137,7 @@ class QuorumPipelineHandler(object):
         """ First split up the main bacteriocin file into a bunch of smaller files"""
         split_bacfiles = ["%s/bacteriocin.%d"%(self.intermediate,i)
                           for i in xrange(njobs)]
-        #self.split_files = split_bacfiles
+        self.split_files += split_bacfiles
         split_bachandles = [open(f,'w') for f in split_bacfiles]
         out_fnames = ["%s/blasted.%d"%(self.intermediate,i) for i in xrange(njobs)]
         out_bac = ["%s.bacteriocins.txt"%(out) for out in out_fnames]
@@ -178,6 +196,7 @@ class QuorumPipelineHandler(object):
                             out_fnames[i])
             
             batch_file = "%s/blast%i.%d.job"%(os.getcwd(),i,os.getpid())
+            self.batch_files.append(batch_file)
             proc = Popen(cmd,shell=True,batch_file=batch_file,
                          stdin=PIPE,stdout=PIPE )
             proc.submit()
@@ -214,6 +233,39 @@ class QuorumPipelineHandler(object):
         self.clusterer.run()
         self.clusterer.parseClusters()
         self.clusterer.dump(self.clusterpickle)
+        
+    """ 
+    Creates HMMER profiles of all of the blasted context genes and the 
+    bacteriocins to find more candidate context genes and bacteriocins
+    """
+    def hmmerGenes(self,min_cluster=10,njobs=2):
+        print "Hmmering Genes"
+        # Aggregate all of the bacteriocins and context genes
+        # into 5 files representing each class 
+        class_handles = [open(c,'w') for c in self.class_files ]
+        for record in SeqIO.parse(self.classifier_out,"fasta"):
+            label = record.id.split('|')[-1]
+            index = self.classes.index(label)
+            SeqIO.write(record,class_handles[index],"fasta")
+        #Move all bacteriocins into toxins file
+        shutil.copyfileobj(open(self.blasted_fasta_bacteriocins),
+                           class_handles[0])
+        for c in class_handles: c.close()
+        #Remove all duplicate ids from this files
+        for fname in self.class_files:
+            f = "%s.%d"%(fname,os.getpid())
+            fasta.remove_duplicates(fname,f)
+            os.rename(f,fname)
+        # Run 6 instances of HMMER for each class
+        self.hmmers = [HMMER(f,quorum,min_cluster) 
+                       for f in self.class_files]
+        for i in xrange(len(self.class_files)):
+            hmmer = self.hmmers[i]
+            hmmer.writeClusters(similarity=0.7,threads=8,memory=3000)
+            hmmer.HMMspawn(njobs)
+            hmmer.search(self.six_frame_genomes,self.class_out[i],njobs)
+
+        
     """ Identifies context genes using BLAST"""
     def blastContextGenes(self,njobs=2):
         print "Blasting Context Genes"
@@ -225,7 +277,7 @@ class QuorumPipelineHandler(object):
         out_classes = ["%s/contextout.%d"%(self.intermediate,i) for i in xrange(njobs)]
         
         index=0
-        for record in SeqIO.parse(self.cand_context_cluster,"fasta"):
+        for record in SeqIO.parse(self.cand_context_genes_fasta,"fasta"):
             split_fastahandles[index].write(">%s\n%s\n"%(str(record.id),
                                                        str(record.seq)))
             index=(index+1)%njobs
@@ -253,6 +305,7 @@ class QuorumPipelineHandler(object):
                                 )
             
             batch_file = "%s/context_blast%i.%d.job"%(os.getcwd(),i,os.getpid())
+            self.batch_files.append(batch_file)
             
             proc = Popen( cmd,shell=True,batch_file=batch_file,
                           stdin=PIPE,stdout=PIPE ) 
@@ -260,19 +313,17 @@ class QuorumPipelineHandler(object):
             #proc.output = out_classes[i]
             jobs.append(proc)
             self.jobs.append(proc)
-        
         for job in jobs: job.wait()
-        
         """ Collect all of the results from the jobs"""
         context_out = open(self.classifier_out,'w')
         for i in xrange(njobs):
             shutil.copyfileobj(open(out_classes[i]),context_out)
-        context_out.close()
-        
+        context_out.close()    
         pass
+    
     """ Classifies individual bacteriocins and context genes based on their text"""
     def textmine(self,njobs=1,dbtype="sqlite3",numTrees=1000):
-        
+        index=0
         print "Classifying"
         """ First split up the main bacteriocin file into a bunch of smaller files"""
         split_fastafiles = ["%s/cluster.%d"%(self.intermediate,i)
@@ -280,8 +331,6 @@ class QuorumPipelineHandler(object):
         self.split_files += split_fastafiles
         split_fastahandles = [open(f,'w') for f in split_fastafiles]
         out_classes = ["%s/classify.%d"%(self.intermediate,i) for i in xrange(njobs)]
-        
-        index=0
         for record in SeqIO.parse(self.cand_context_cluster,"fasta"):
             split_fastahandles[index].write(">%s\n%s\n"%(str(record.id),
                                                        str(record.seq)))
@@ -314,7 +363,7 @@ class QuorumPipelineHandler(object):
                                 )
             
             batch_file = "%s/classify%i.%d.job"%(os.getcwd(),i,os.getpid())
-            
+            self.batch_files.append(batch_file)
             proc = Popen( cmd,shell=True,batch_file=batch_file,
                           stdin=PIPE,stdout=PIPE ) 
             proc.submit()
@@ -337,13 +386,16 @@ if __name__=="__main__":
         'Finds bacteriocins and context genes')
     parser.add_argument(\
         '--pipeline-section', type=str, required=False, default="all",
-        help='Section of the pipeline to run (all, preprocess, blast, cluster, nbayes, mvote)')
+        help='Section of the pipeline to run (all, preprocess, blast, context, hmmer)')
     parser.add_argument(\
         '--root-dir',type=str, required=False,
         help='Root directory')
     parser.add_argument(\
         '--genome-files', type=str, nargs="+", required=False,
         help='FASTA files containing bacterial genomes')
+    parser.add_argument(\
+        '--six-frame-genome', type=str, required=False,
+        help='FASTA file containing six-frame translated bacterial genomes')
     parser.add_argument(\
         '--intergenes', type=str, required=False,default=None,
         help='FASTA files containing intergenic regions')
@@ -357,8 +409,11 @@ if __name__=="__main__":
         '--bacteriocin-radius', type=int, required=False, default=5000,
         help='The search radius around every specified bacteriocin')
     parser.add_argument(\
-        '--similarity', type=int, required=False, default=0.7,
+        '--similarity', type=float, required=False, default=0.7,
         help='Clustering similarity')    
+    parser.add_argument(\
+        '--cluster-size', type=int, required=False, default=10,
+        help='Filters all clusters below this threshold')    
     parser.add_argument(\
         '--bac-evalue', type=float, required=False, default=0.00001,
         help='The evalue for bacteriocin hits')
@@ -400,6 +455,7 @@ if __name__=="__main__":
         print "Formatdb:",args.formatdb
         proc = QuorumPipelineHandler(args.root_dir,
                                      args.genome_files,
+                                     args.six_frame_genome,
                                      args.intergenes,
                                      args.annotated_genes,
                                      args.bacteriocins,
@@ -415,9 +471,15 @@ if __name__=="__main__":
                                      args.formatdb,
                                      args.verbose                
                                     ) 
-        proc.blast(njobs=args.num_jobs)
-        proc.cluster()
-        proc.blastContextGenes(njobs=args.num_jobs)
+        if args.pipeline_section=="blast":
+            proc.blast(njobs=args.num_jobs)
+            proc.blastContextGenes(njobs=args.num_jobs)
+            proc.hmmerGenes(args.cluster_size,args.num_jobs)
+        elif args.pipeline_section=="context":
+            proc.blastContextGenes(njobs=args.num_jobs)
+            proc.hmmerGenes(args.cluster_size,args.num_jobs)
+        else:
+            proc.hmmerGenes(args.cluster_size,args.num_jobs)
     else:
         del sys.argv[1:]
         import unittest
@@ -427,7 +489,8 @@ if __name__=="__main__":
                 self.root = os.environ['BACFINDER_HOME']
                 self.exampledir = "%s/example/Streptococcus_pyogenes"%self.root
                 self.bacdir = "%s/bacteriocins"%self.root
-                self.genome_files = test_modules.getFNA(self.exampledir)
+                self.genome_files = test_modules.getFNA(self.exampledir) 
+                self.six_frame_genome = "%s/example/all_trans.fna"%self.root
                 self.bacteriocins = "%s/bagel.fa"%self.bacdir
                 self.intergenes = "test_intergenes.fa"
                 self.annotated_genes = "test_genes.fa"
@@ -448,18 +511,19 @@ if __name__=="__main__":
                 self.verbose = True
                 self.keep_tmp = False
                 self.similarity = 0.65
-                self.numThreads = 1
+                self.numThreads = 8
                 self.output = "out"
                 self.keep_tmp = True
                 self.proc = None
             def tearDown(self):
                 self.proc.cleanup()
-                pass
+                
             def testrun(self):
                 print "Test Run"
                 self.proc = QuorumPipelineHandler(  
                                          self.root,
                                          self.genome_files,
+                                         self.six_frame_genome,
                                          self.intergenes,
                                          self.annotated_genes,
                                          self.bacteriocins,
@@ -480,19 +544,23 @@ if __name__=="__main__":
                 self.assertTrue(os.path.getsize(self.annotated_genes) > 0)
                 self.assertTrue(os.path.getsize(self.intergenes) > 0)
                 
-                self.proc.blast(njobs=2)
+                self.proc.blast(njobs=3)
                 
                 self.assertTrue(os.path.getsize(self.proc.blasted_fasta_bacteriocins) > 0)
                 self.assertTrue(os.path.getsize(self.proc.cand_context_genes_fasta) > 0)
                 
-                self.proc.cluster()
-                
-                self.assertTrue(os.path.getsize("%s"%(self.proc.cand_context_genes_fasta)) > 0)                
-                self.assertTrue(os.path.getsize(self.proc.cand_context_cluster) > 0)
-                
+                #self.proc.cluster()
+                #self.assertTrue(os.path.getsize("%s"%(self.proc.cand_context_genes_fasta)) > 0)                
+                #self.assertTrue(os.path.getsize(self.proc.cand_context_cluster) > 0)
+                                
                 #self.proc.textmine(njobs=7)
-                self.proc.blastContextGenes()
+                self.proc.blastContextGenes(njobs=3)
                 self.assertTrue(os.path.getsize( self.proc.classifier_out ) > 0)
+                
+                self.proc.hmmerGenes(min_cluster=60,njobs=10)
+                for fname in self.proc.class_files:
+                    self.assertTrue(os.path.getsize(fname)>0)
+                
                 #self.proc.cleanup()
                 
         unittest.main()
